chore: Extract RecordSchema into its own module (src/schema/record/schema.rs)
